backend/requirements.txt
asgiref==3.7.2
distlib==0.3.4
Django==4.2.3
djangorestframework==3.14.0
filelock==3.7.0
platformdirs==2.5.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
virtualenv==20.14.1
